Implementing LongTask with requestIdleCallback
This makes all LongTasks truely async which gives browser a chance to
update the UI before and after task is run. There's a good chance we'll
need to add a rIC polyfill for wider browser support. The polyfill seems
pretty reasonable, it's essentially a call to setTimeout.

Also cleaned up LongTask implementation a little bit, added AsyncTask to
allow for configurable promise implementations.